Squashed 'airdcpp-core/' changes from 9ce2e9bc8..d86d1f01d
d86d1f01d Refactor updater and startup params
bec28c020 Fix a few quality warnings from Sonarcloud
511b34f26 Filelist-related bug and crash fixes
10d97d581 Fix a few compiler warnings with the 32 bit build
ca6144a23 Move third-party libraries into a separate directory

git-subtree-dir: airdcpp-core
git-subtree-split: d86d1f01dfdb6a6e47f8bc3a2ca82e12320197d3

+++ airdcpp/Hasher.cpp
@@ -214,7 +214,7 @@ void Hasher::getStats(string& curFile_, int64_t& bytesLeft_, size_t& filesLeft_,
 
 	bytesLeft_ += totalBytesLeft;
 
-	filesAdded_ += totalFilesAdded;
+	filesAdded_ += static_cast<size_t>(totalFilesAdded);
 	bytesAdded_ += totalBytesAdded;
 }
